+++ esp-hal-procmacros/src/ram.rs
@@ -10,6 +10,7 @@ pub fn ram(args: TokenStream, input: TokenStream) -> TokenStream {
 
     let mut rtc_fast = false;
     let mut rtc_slow = false;
+    let mut dram2_uninit = false;
     let mut persistent = false;
     let mut zeroed = false;
 
@@ -22,6 +23,7 @@ pub fn ram(args: TokenStream, input: TokenStream) -> TokenStream {
             let arg = match ident {
                 i if i == "rtc_fast" => &mut rtc_fast,
                 i if i == "rtc_slow" => &mut rtc_slow,
+                i if i == "reclaimed" => &mut dram2_uninit,
                 i if i == "persistent" => &mut persistent,
                 i if i == "zeroed" => &mut zeroed,
                 i => {
@@ -53,20 +55,21 @@ pub fn ram(args: TokenStream, input: TokenStream) -> TokenStream {
     }
 
     let is_fn = matches!(item, Item::Fn(_));
-    let section_name = match (is_fn, rtc_fast, rtc_slow, persistent, zeroed) {
-        (true, false, false, false, false) => Ok(".rwtext"),
-        (true, true, false, false, false) => Ok(".rtc_fast.text"),
-        (true, false, true, false, false) => Ok(".rtc_slow.text"),
+    let section_name = match (is_fn, rtc_fast, rtc_slow, dram2_uninit, persistent, zeroed) {
+        (true, false, false, false, false, false) => Ok(".rwtext"),
+        (true, true, false, false, false, false) => Ok(".rtc_fast.text"),
+        (true, false, true, false, false, false) => Ok(".rtc_slow.text"),
 
-        (false, false, false, false, false) => Ok(".data"),
+        (false, false, false, false, false, false) => Ok(".data"),
+        (false, false, false, true, false, false) => Ok(".dram2_uninit"),
 
-        (false, true, false, false, false) => Ok(".rtc_fast.data"),
-        (false, true, false, true, false) => Ok(".rtc_fast.persistent"),
-        (false, true, false, false, true) => Ok(".rtc_fast.bss"),
+        (false, true, false, false, false, false) => Ok(".rtc_fast.data"),
+        (false, true, false, false, true, false) => Ok(".rtc_fast.persistent"),
+        (false, true, false, false, false, true) => Ok(".rtc_fast.bss"),
 
-        (false, false, true, false, false) => Ok(".rtc_slow.data"),
-        (false, false, true, true, false) => Ok(".rtc_slow.persistent"),
-        (false, false, true, false, true) => Ok(".rtc_slow.bss"),
+        (false, false, true, false, false, false) => Ok(".rtc_slow.data"),
+        (false, false, true, false, true, false) => Ok(".rtc_slow.persistent"),
+        (false, false, true, false, false, true) => Ok(".rtc_slow.bss"),
 
         _ => Err(()),
     };
@@ -90,6 +93,8 @@ pub fn ram(args: TokenStream, input: TokenStream) -> TokenStream {
         Some("zeroable")
     } else if persistent {
         Some("persistable")
+    } else if dram2_uninit {
+        Some("uninit")
     } else {
         None
     };